backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
joblib==1.4.2
mongoengine==0.28.2
mongomock==4.1.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
setuptools==75.1.0
